refactor: isolate qihe actor
